requirements.txt
aiofiles
aioshutil
python-dotenv
pygit2
packaging
aiohttp
pip
msgpack
anyio
jurigged
thefuzz

hikari
hikari-arc
hikari-miru

ciso8601
orjson
uvloop
speedups

lmdb